Siren Komik: fix NPE
