+++ src/id/mangkomik/build.gradle
@@ -1,9 +1,9 @@
 ext {
     extName = 'Siren Komik'
-    extClass = '.MangKomik'
+    extClass = '.SirenKomik'
     themePkg = 'mangathemesia'
     baseUrl = 'https://sirenkomik.my.id'
-    overrideVersionCode = 2
+    overrideVersionCode = 3
 }
 
 apply from: "$rootDir/common.gradle"
